bring back the legacy data table
